Separating functilnality of processes in utils, got ingest XML working <response></response> was the issue so skipped it and set root of XML tree to it's first element [0]
